requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
appdirs=1.4.4=pypi_0
blas=1.0=mkl
bzip2=1.0.8=h5eee18b_5
ca-certificates=2023.12.12=h06a4308_0
certifi=2024.2.2=py310h06a4308_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.7=pypi_0
contourpy=1.2.0=pypi_0
cuda-cudart=11.8.89=0
cuda-cupti=11.8.87=0
cuda-libraries=11.8.0=0
cuda-nvrtc=11.8.89=0
cuda-nvtx=11.8.86=0
cuda-runtime=11.8.0=0
cycler=0.12.1=pypi_0
docker-pycreds=0.4.0=pypi_0
einops=0.7.0=pypi_0
ffmpeg=4.3=hf484d3e_0
filelock=3.13.1=py310h06a4308_0
fonttools=4.50.0=pypi_0
freetype=2.12.1=h4a9f257_0
ftfy=6.1.3=pypi_0
gitdb=4.0.11=pypi_0
gitpython=3.1.42=pypi_0
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py310heeb90bb_0
gnutls=3.6.15=he1e5248_0
idna=3.4=py310h06a4308_0
intel-openmp=2023.1.0=hdb19cb5_46306
jinja2=3.1.3=py310h06a4308_0
joblib=1.2.0=py310h06a4308_0
jpeg=9e=h5eee18b_1
kiwisolver=1.4.5=pypi_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libcublas=11.11.3.6=0
libcufft=10.9.0.58=0
libcufile=1.9.0.20=0
libcurand=10.3.5.119=0
libcusolver=11.4.1.48=0
libcusparse=11.7.5.86=0
libdeflate=1.8=h7f8727e_5
libffi=3.4.2=h6a678d5_6
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.4=h5eee18b_0
libjpeg-turbo=2.0.0=h9bf148f_0
libnpp=11.8.0.86=0
libnvjpeg=11.9.0.86=0
libpng=1.6.39=h5eee18b_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.0=hecacb30_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libwebp-base=1.3.2=h5eee18b_0
lightning-utilities=0.9.0=py310h06a4308_0
llvm-openmp=14.0.6=h9e868ea_0
lz4-c=1.9.4=h6a678d5_0
markupsafe=2.1.3=py310h5eee18b_0
matplotlib=3.8.3=pypi_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py310h5eee18b_1
mkl_fft=1.3.8=py310h5eee18b_0
mkl_random=1.2.4=py310hdb19cb5_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.3.0=py310h06a4308_0
ncurses=6.3=h5eee18b_3
nettle=3.7.3=hbbd107a_1
networkx=3.1=py310h06a4308_0
numpy=1.26.4=py310h5f9d8c6_0
numpy-base=1.26.4=py310hb5e798b_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=h3ad879b_0
openssl=1.1.1w=h7f8727e_0
packaging=23.1=py310h06a4308_0
pillow=10.2.0=py310h5eee18b_0
pip=22.2.2=py310h06a4308_0
protobuf=4.25.3=pypi_0
psutil=5.9.8=pypi_0
pyparsing=3.1.2=pypi_0
python=3.10.8=h7a1cb2a_1
python-dateutil=2.9.0.post0=pypi_0
pytorch=2.2.1=py3.10_cuda11.8_cudnn8.7.0_0
pytorch-cuda=11.8=h7e8668a_5
pytorch-mutex=1.0=cuda
pyyaml=6.0.1=py310h5eee18b_0
readline=8.2=h5eee18b_0
regex=2023.12.25=pypi_0
requests=2.31.0=py310h06a4308_1
scikit-learn=1.3.0=py310h1128e8f_1
scipy=1.11.4=py310h5f9d8c6_0
sentry-sdk=1.42.0=pypi_0
setproctitle=1.3.3=pypi_0
setuptools=65.5.0=py310h06a4308_0
six=1.16.0=pypi_0
smmap=5.0.1=pypi_0
sqlite=3.40.0=h5082296_0
sympy=1.12=py310h06a4308_0
tbb=2021.8.0=hdb19cb5_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.12=h1ccaba5_0
torchaudio=2.2.1=py310_cu118
torchmetrics=1.1.2=py310h06a4308_0
torchtriton=2.2.0=py310
torchvision=0.17.1=py310_cu118
tqdm=4.66.2=pypi_0
typing_extensions=4.9.0=py310h06a4308_1
tzdata=2024a=h04d1e81_0
urllib3=2.1.0=py310h06a4308_0
wandb=0.16.4=pypi_0
wcwidth=0.2.13=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.8=h5eee18b_0
yaml=0.2.5=h7b6447c_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.2=ha4553b6_0
